Final pop animation commit
Tiles now succesfully pop out in the foremost layer by delaying the render command to let it render in front of the other tiles. Tiles also use a 'size' command to immediately create contrast between popped and unpopped tiles, and to tell the game when to unrender the tile (TBD). The next issue to be fixed is that the frame rate issues return when all of the tiles have been uncovered,
- Created a new int 'size'
- Changed the resize command to depend of size
- FIxed an issue where resize would freak out when size is less then zero
- Created a new interative loop for the popped tiles.